Refactoring and getting tests to work with latest DB changes (UUID)
